The revisions button is a permanent member of the global styles sidebar, so it doesn't need to be included via fill. This commit moves the button from ui.js to the sidebar component in the hope that it will make both maintenance and working with parallel states, e.g., style book visibility easier later.
